support urlencoded form data post from gurl. add logger to curl adapter add stub scrape method to jss add scrape test
