+++ tests/conftest.py
@@ -100,6 +100,18 @@ def dp_smb_ip_port(docker_ip, docker_services):
     return docker_ip, docker_services.port_for('samba', 139)
 
 
+@pytest.fixture
+def gurl_jss(gurl_adapter, jss_prefs_dict):  # type: (GurlAdapter, dict) -> JSS
+    j = JSS(
+        adapter=gurl_adapter,
+        url=jss_prefs_dict['jss_url'],
+        user=jss_prefs_dict['jss_user'],
+        password=jss_prefs_dict['jss_password'],
+        ssl_verify=jss_prefs_dict['verify'],
+    )
+    return j
+
+
 # @pytest.fixture
 # def dp_afp_url(docker_ip, docker_services):
 #     afp_url = 'afp://%s:%s/distribution_point' % (
