+++ tests/test_gurl_adapter.py
@@ -5,16 +5,6 @@ from jss import JSS, Building
 from xml.etree import ElementTree
 
 
-@pytest.fixture
-def gurl_jss(gurl_adapter, jss_prefs_dict):  # type: (GurlAdapter, dict) -> JSS
-    j = JSS(
-        adapter=gurl_adapter,
-        url=jss_prefs_dict['jss_url'],
-        user=jss_prefs_dict['jss_user'],
-        password=jss_prefs_dict['jss_password'],
-        ssl_verify=jss_prefs_dict['verify'],
-    )
-    return j
 
 
 @pytest.mark.skipif(sys.platform.startswith('linux'), reason='PyObjC not present on linux')
